refactor: moved all aql statement into its own package
